Crate wgpu_hal

This library describes the internal unsafe graphics abstraction API. It follows WebGPU for the most part, re-using wgpu-types, with the following deviations:

Fully unsafe: zero overhead, zero validation.

Compile-time backend selection via traits.

Objects are passed by references and returned by value. No IDs.

Mapping is persistent, with explicit synchronization.

Resource transitions are explicit.

All layouts are explicit. Binding model has compatibility.

General design direction is to follow the majority by the following weights:

wgpu-core: 1.5

primary backends (Vulkan/Metal/DX12): 1.0 each

secondary backend (GLES): 0.5

Modules

api

auxil

empty

A dummy API implementation.

gles

GLES API internals.

vulkan

Vulkan API internals.

Structs

AccelerationStructureAABBs

offset - offset in bytes

AccelerationStructureBarrier

AccelerationStructureBuildFlags

Flags for acceleration structures

AccelerationStructureBuildSizes

Information of the required size for a corresponding entries struct (+ flags)

AccelerationStructureDescriptor

AccelerationStructureGeometryFlags

Flags for acceleration structure geometries

AccelerationStructureInstances

offset - offset in bytes

AccelerationStructureTriangleIndices

offset - offset in bytes

AccelerationStructureTriangleTransform

offset - offset in bytes

AccelerationStructureTriangles

first_vertex - offset in the vertex buffer (as number of vertices)indices - optional index buffer with attributestransform - optional transform

AccelerationStructureUses

AcquiredSurfaceTexture

Alignments

Attachment

AttachmentOps

BindGroupDescriptor

BindGroup descriptor.

BindGroupEntry

BindGroupLayoutDescriptor

BindGroupLayout descriptor.

BindGroupLayoutFlags

Pipeline layout creation flags.

BufferBarrier

BufferBinding

BufferCopy

BufferDescriptor

BufferMapping

BufferTextureCopy

BufferUses

Similar to wgt::BufferUsages but for internal use.

BuildAccelerationStructureDescriptor

Updates use source_acceleration_structure if present, else the update will be performed in place. For updates, only the data is allowed to change (not the meta data or sizes).

Capabilities

ColorAttachment

CommandEncoderDescriptor

ComputePassDescriptor

ComputePassTimestampWrites

ComputePipelineDescriptor

Describes a compute pipeline.

CopyExtent

DebugSource

DepthStencilAttachment

ExposedAdapter

FormatAspects

Texture format capability flags.

GetAccelerationStructureBuildSizesDescriptor

All buffers, buffer addresses and offsets will be ignored.The build mode will be ignored.Reducing the amount of Instances, Triangle groups or AABB groups (or the number of Triangles/AABBs in corresponding groups), may result in reduced size requirements.Any other change may result in a bigger or smaller size requirement.

InstanceDescriptor

InstanceError

Error occurring while trying to create an instance, or create a surface from an instance; typically relating to the state of the underlying graphics API or hardware.

MemoryFlags

NagaShader

Naga shader module.

OpenDevice

PipelineLayoutDescriptor

PipelineLayoutFlags

Pipeline layout creation flags.

ProgrammableStage

Describes a programmable pipeline stage.

Rect

RenderPassDescriptor

RenderPassTimestampWrites

RenderPipelineDescriptor

Describes a render (graphics) pipeline.

SamplerDescriptor

ShaderModuleDescriptor

SurfaceCapabilities

Describes information about what a Surface’s presentation capabilities are. Fetch this with Adapter::surface_capabilities.

SurfaceConfiguration

TextureBarrier

TextureBinding

TextureCopy

TextureCopyBase

TextureDescriptor

TextureFormatCapabilities

Texture format capability flags.

TextureUses

Similar to wgt::TextureUsages but for internal use.

TextureViewDescriptor

TextureView descriptor.

ValidationCanary

Flag for internal testing.

VertexBufferLayout

Describes how the vertex buffer is interpreted.

Enums

AccelerationStructureBuildMode

AccelerationStructureEntries

Entries for a single descriptor

AccelerationStructureFormat

DeviceError

PipelineError

ShaderError

ShaderInput

Shader input.

SurfaceError

Constants

MAX_ANISOTROPY

MAX_BIND_GROUPS

MAX_COLOR_ATTACHMENTS

MAX_CONCURRENT_SHADER_STAGES

MAX_MIP_LEVELS

MAX_VERTEX_BUFFERS

QUERY_SIZE

Size of a single occlusion/timestamp query, when copied into a buffer, in bytes.

Statics

VALIDATION_CANARY

Stores the text of any validation errors that have occurred since the last call to get_and_reset.

Traits

Adapter

Api

CommandEncoder

Encoder for commands in command buffers. Serves as a parent for all the encoded command buffers. Works in bursts of action: one or more command buffers are recorded, then submitted to a queue, and then it needs to be reset_all().

Device

Instance

Queue

Surface

Type Aliases

DropGuard

Drop guard to signal wgpu-hal is no longer using an externally created object.

FenceValue

Label

MemoryRange