wgpu_hal::gles

Struct Adapter

pub struct Adapter {
    shared: Arc<AdapterShared>,
}

Available on gles only.

Fields
shared: Arc<AdapterShared>
Implementations

impl Adapter

pub unsafe fn new_external( fun: impl FnMut(&str) -> *const c_void, options: GlBackendOptions, ) -> Option<ExposedAdapter<Api>>

Available on neither Windows nor webgl.

Creates a new external adapter using the specified loader function.

Safety

The underlying OpenGL ES context must be current.

The underlying OpenGL ES context must be current when interfacing with any objects returned by wgpu-hal from this adapter.

The underlying OpenGL ES context must be current when dropping this adapter and when dropping any objects returned from this adapter.

pub fn adapter_context(&self) -> &AdapterContext

Available on neither Windows nor webgl.

impl Adapter

fn parse_version(src: &str) -> Result<(u8, u8), InstanceError>

Note that this function is intentionally lenient in regards to parsing, and will try to recover at least the first two version numbers without resulting in an Err.

Notes

WebGL 2 version returned as OpenGL ES 3.0

pub(super) fn parse_full_version(src: &str) -> Result<(u8, u8), InstanceError>

According to the OpenGL specification, the version information is expected to follow the following syntax:

<major>       ::= <number>
<minor>       ::= <number>
<revision>    ::= <number>
<vendor-info> ::= <string>
<release>     ::= <major> "." <minor> ["." <release>]
<version>     ::= <release> [" " <vendor-info>]

Note that this function is intentionally lenient in regards to parsing, and will try to recover at least the first two version numbers without resulting in an Err.

fn make_info( vendor_orig: String, renderer_orig: String, version: String, ) -> AdapterInfo

pub(super) unsafe fn expose( context: AdapterContext, backend_options: GlBackendOptions, ) -> Option<ExposedAdapter<Api>>

unsafe fn compile_shader( source: &str, gl: &Context, shader_type: u32, es: bool, ) -> Option<Shader>

unsafe fn create_shader_clear_program( gl: &Context, es: bool, ) -> Option<ShaderClearProgram>