wgpu_hal::gles
Struct AdapterShared

struct AdapterShared {
    context: AdapterContext,
    private_caps: PrivateCapabilities,
    features: Features,
    limits: Limits,
    workarounds: Workarounds,
    options: GlBackendOptions,
    shading_language_version: Version,
    next_shader_id: AtomicU32,
    program_cache: Mutex<FastHashMap<ProgramCacheKey, Result<Arc<PipelineInner>, PipelineError>>>,
    es: bool,
    max_msaa_samples: i32,
}

Available on gles only.

Fields
context: AdapterContextprivate_caps: PrivateCapabilitiesfeatures: Featureslimits: Limitsworkarounds: Workaroundsoptions: GlBackendOptionsshading_language_version: Versionnext_shader_id: AtomicU32program_cache: Mutex<FastHashMap<ProgramCacheKey, Result<Arc<PipelineInner>, PipelineError>>>es: boolmax_msaa_samples: i32

Result of gl.get_parameter_i32(glow::MAX_SAMPLES). Cached here so it doesn’t need to be queried every time texture format capabilities are requested. (this has been shown to be a significant enough overhead)

Implementations

impl AdapterShared

pub(super) unsafe fn get_buffer_sub_data( &self, gl: &Context, target: u32, offset: i32, dst_data: &mut [u8], )

impl AdapterShared

pub(super) fn describe_texture_format( &self, texture_format: TextureFormat, ) -> TextureFormatDesc