wgpu_hal::gles
Struct Queue

pub struct Queue {
    shared: Arc<AdapterShared>,
    features: Features,
    draw_fbo: Framebuffer,
    copy_fbo: Framebuffer,
    shader_clear_program: Option<ShaderClearProgram>,
    zero_buffer: Buffer,
    temp_query_results: Mutex<Vec<u64>>,
    draw_buffer_count: AtomicU8,
    current_index_buffer: Mutex<Option<Buffer>>,
}

Available on gles only.

Fields
shared: Arc<AdapterShared>features: Featuresdraw_fbo: Framebuffercopy_fbo: Framebuffershader_clear_program: Option<ShaderClearProgram>

Shader program used to clear the screen for Workarounds::MESA_I915_SRGB_SHADER_CLEAR devices.

zero_buffer: Buffer

Keep a reasonably large buffer filled with zeroes, so that we can implement ClearBuffer of zeroes by copying from it.

temp_query_results: Mutex<Vec<u64>>draw_buffer_count: AtomicU8current_index_buffer: Mutex<Option<Buffer>>
Implementations

impl Queue

unsafe fn perform_shader_clear( &self, gl: &Context, draw_buffer: u32, color: [f32; 4], )

Performs a manual shader clear, used as a workaround for a clearing bug on mesa

unsafe fn reset_state(&self, gl: &Context)

unsafe fn set_attachment( &self, gl: &Context, fbo_target: u32, attachment: u32, view: &TextureView, )

unsafe fn process( &self, gl: &Context, command: &C, data_bytes: &[u8], queries: &[Query], )