wgpu_hal
Module metal

Metal API internals.

Metal API internals.

Pipeline Layout

In Metal, push constants, vertex buffers, and resources in the bind groups are all placed together in the native resource bindings, which work similarly to D3D11: there are tables of textures, buffers, and samplers.

We put push constants first (if any) in the table, followed by bind group 0 resources, followed by other bind groups. The vertex buffers are bound at the very end of the VS buffer table.

!

Structs

Adapter

Api

BindGroup

BindGroupLayout

Buffer

CommandBuffer

CommandEncoder

ComputePipeline

Device

Fence

Instance

PipelineLayout

QuerySet

Queue

RenderPipeline

Sampler

ShaderModule

Surface

SurfaceTexture

Texture

TextureView