wgpu_hal::noop::command
Struct CommandBuffer

pub struct CommandBuffer {
    commands: Vec<Command>,
}

Command buffer type, which performs double duty as the command encoder type too.

Fields
commands: Vec<Command>
Implementations

impl CommandBuffer

pub(crate) unsafe fn execute(&self)

Safety

Must be called with appropriate synchronization for the resources affected by the command, such as ensuring that buffers are not accessed by a command while aliasing references exist.

pub(crate) fn new() -> Self