Struct wgpu_hal::vulkan::adapter::PhysicalDeviceFeatures

pub struct PhysicalDeviceFeatures {
    core: PhysicalDeviceFeatures,
    pub(super) descriptor_indexing: Option<PhysicalDeviceDescriptorIndexingFeaturesEXT>,
    imageless_framebuffer: Option<PhysicalDeviceImagelessFramebufferFeaturesKHR>,
    timeline_semaphore: Option<PhysicalDeviceTimelineSemaphoreFeaturesKHR>,
    image_robustness: Option<PhysicalDeviceImageRobustnessFeaturesEXT>,
    robustness2: Option<PhysicalDeviceRobustness2FeaturesEXT>,
    multiview: Option<PhysicalDeviceMultiviewFeaturesKHR>,
    astc_hdr: Option<PhysicalDeviceTextureCompressionASTCHDRFeaturesEXT>,
    shader_float16: Option<(PhysicalDeviceShaderFloat16Int8Features, PhysicalDevice16BitStorageFeatures)>,
    zero_initialize_workgroup_memory: Option<PhysicalDeviceZeroInitializeWorkgroupMemoryFeatures>,
}

Aggregate of the vk::PhysicalDevice*Features structs used by gfx.

Fields
core: PhysicalDeviceFeaturesdescriptor_indexing: Option<PhysicalDeviceDescriptorIndexingFeaturesEXT>imageless_framebuffer: Option<PhysicalDeviceImagelessFramebufferFeaturesKHR>timeline_semaphore: Option<PhysicalDeviceTimelineSemaphoreFeaturesKHR>image_robustness: Option<PhysicalDeviceImageRobustnessFeaturesEXT>robustness2: Option<PhysicalDeviceRobustness2FeaturesEXT>multiview: Option<PhysicalDeviceMultiviewFeaturesKHR>astc_hdr: Option<PhysicalDeviceTextureCompressionASTCHDRFeaturesEXT>shader_float16: Option<(PhysicalDeviceShaderFloat16Int8Features, PhysicalDevice16BitStorageFeatures)>zero_initialize_workgroup_memory: Option<PhysicalDeviceZeroInitializeWorkgroupMemoryFeatures>
Implementations

impl PhysicalDeviceFeatures

pub fn add_to_device_create_builder<'a>( &'a mut self, info: DeviceCreateInfoBuilder<'a> ) -> DeviceCreateInfoBuilder<'a>

Add the members of self into info.enabled_features and its p_next chain.

fn from_extensions_and_requested_features( device_api_version: u32, enabled_extensions: &[&'static CStr], requested_features: Features, downlevel_flags: DownlevelFlags, private_caps: &PrivateCapabilities ) -> Self

Create a PhysicalDeviceFeatures that will be used to create a logical device.

requested_features should be the same as what was used to generate enabled_extensions.

fn to_wgpu( &self, instance: &Instance, phd: PhysicalDevice, caps: &PhysicalDeviceCapabilities ) -> (Features, DownlevelFlags)

fn all_features_supported( features: &Features, implications: &[(Features, Bool32)] ) -> bool