wgpu_hal::vulkan

Module descriptor

Available on vulkan only.

Structs

Bucket 🔒

Keeps track of all pools created with this bucket’s BucketKey.

BucketKey 🔒

DescriptorAllocator

DescriptorCounts

The dynamic counts are tracked separately since dynamic textures/buffers have distinct descriptor types compared to their non-dynamic counterparts.

DescriptorSet

Pool 🔒

Constants

POOL_MAX_SETS 🔒

POOL_MIN_SETS 🔒

Functions

create_descriptor_pool 🔒