wgpu_hal::vulkan
Module device

Available on vulkan only.

Structs

CompiledStage 🔒

ImageWithoutMemory 🔒

Functions

handle_unexpected 🔒

We usually map unexpected vulkan errors to the crate::DeviceError::Unexpected variant to be more robust even in cases where the driver is not complying with the spec.