wgpu_hal::vulkan::device

Function handle_unexpected

fn handle_unexpected(err: Result) -> !

Available on vulkan only.

We usually map unexpected vulkan errors to the crate::DeviceError::Unexpected variant to be more robust even in cases where the driver is not complying with the spec.

However, we implement a few Trait methods that don’t have an equivalent error variant. In those cases we use this function.