wgpu_hal::vulkan

Struct Adapter

pub struct Adapter {
    raw: PhysicalDevice,
    instance: Arc<InstanceShared>,
    known_memory_flags: MemoryPropertyFlags,
    phd_capabilities: PhysicalDeviceProperties,
    phd_features: PhysicalDeviceFeatures,
    downlevel_flags: DownlevelFlags,
    private_caps: PrivateCapabilities,
    workarounds: Workarounds,
}

Fields
raw: PhysicalDeviceinstance: Arc<InstanceShared>known_memory_flags: MemoryPropertyFlagsphd_capabilities: PhysicalDevicePropertiesphd_features: PhysicalDeviceFeaturesdownlevel_flags: DownlevelFlagsprivate_caps: PrivateCapabilitiesworkarounds: Workarounds
Implementations

impl Adapter

pub fn raw_physical_device(&self) -> PhysicalDevice

pub fn get_physical_device_features(&self) -> &PhysicalDeviceFeatures

pub fn physical_device_capabilities(&self) -> &PhysicalDeviceProperties

pub fn shared_instance(&self) -> &InstanceShared

pub fn required_device_extensions( &self, features: Features, ) -> Vec<&'static CStr>

pub fn physical_device_features( &self, enabled_extensions: &[&'static CStr], features: Features, ) -> PhysicalDeviceFeatures

Create a PhysicalDeviceFeatures for opening a logical device with features from this adapter.

The given enabled_extensions set must include all the extensions selected by required_device_extensions when passed features. Otherwise, the PhysicalDeviceFeatures value may not be able to select all the Vulkan features needed to represent features and this adapter’s characteristics.

Typically, you’d simply call required_device_extensions, and then pass its return value and the feature set you gave it directly to this function. But it’s fine to add more extensions to the list.

pub unsafe fn device_from_raw( &self, raw_device: Device, drop_callback: Option<DropCallback>, enabled_extensions: &[&'static CStr], features: Features, memory_hints: &MemoryHints, family_index: u32, queue_index: u32, ) -> Result<OpenDevice<Api>, DeviceError>

Safety

raw_device must be created from this adapter.

raw_device must be created using family_index, enabled_extensions and physical_device_features()

enabled_extensions must be a superset of required_device_extensions().

If drop_callback is None, wgpu-hal will take ownership of raw_device. If drop_callback is Some, raw_device must be valid until the callback is called.

pub fn texture_format_as_raw(&self, texture_format: TextureFormat) -> Format

pub unsafe fn open_with_callback<'a>( &self, features: Features, memory_hints: &MemoryHints, callback: Option<Box<CreateDeviceCallback<'a>>>, ) -> Result<OpenDevice<Api>, DeviceError>

Safety:

Same as open plus

The callback may not change anything that the device does not support.

The callback may not remove features.