wgpu_hal::vulkan
Struct CommandEncoder

pub struct CommandEncoder {
    raw: CommandPool,
    device: Arc<DeviceShared>,
    active: CommandBuffer,
    bind_point: PipelineBindPoint,
    temp: Temp,
    free: Vec<CommandBuffer>,
    discarded: Vec<CommandBuffer>,
    rpass_debug_marker_active: bool,
    end_of_pass_timer_query: Option<(QueryPool, u32)>,
    counters: Arc<HalCounters>,
}

Available on vulkan only.

Fields
raw: CommandPooldevice: Arc<DeviceShared>active: CommandBuffer

The current command buffer, if self is in the “recording” state.

If non-null, the buffer is in the Vulkan “recording” state.

bind_point: PipelineBindPoint

What kind of pass we are currently within: compute or render.

temp: Temp

Allocation recycling pool for this encoder.

free: Vec<CommandBuffer>

A pool of available command buffers.

These are all in the Vulkan “initial” state.

discarded: Vec<CommandBuffer>

A pool of discarded command buffers.

These could be in any Vulkan state except “pending”.

rpass_debug_marker_active: bool

If this is true, the active renderpass enabled a debug span, and needs to be disabled on renderpass close.

end_of_pass_timer_query: Option<(QueryPool, u32)>

If set, the end of the next render/compute pass will write a timestamp at the given pool & location.

counters: Arc<HalCounters>
Implementations

impl CommandEncoder

fn write_pass_end_timestamp_if_requested(&mut self)

impl CommandEncoder

pub unsafe fn raw_handle(&self) -> CommandBuffer

Safety

The command buffer handle must not be manually destroyed