Struct wgpu_hal::vulkan::Device

pub struct Device {
    shared: Arc<DeviceShared>,
    mem_allocator: Mutex<GpuAllocator<DeviceMemory>>,
    desc_allocator: Mutex<DescriptorAllocator<DescriptorPool, DescriptorSet>>,
    valid_ash_memory_types: u32,
    naga_options: Options<'static>,
}

Fields
shared: Arc<DeviceShared>mem_allocator: Mutex<GpuAllocator<DeviceMemory>>desc_allocator: Mutex<DescriptorAllocator<DescriptorPool, DescriptorSet>>valid_ash_memory_types: u32naga_options: Options<'static>
Implementations

impl Device

pub(super) unsafe fn create_swapchain( &self, surface: &mut Surface, config: &SurfaceConfiguration, provided_old_swapchain: Option<Swapchain> ) -> Result<Swapchain, SurfaceError>

pub unsafe fn texture_from_raw( vk_image: Image, desc: &TextureDescriptor<'_>, drop_guard: Option<DropGuard> ) -> Texture

Safety

vk_image must be created respecting desc

If drop_guard is Some, the application must manually destroy the image handle. This can be done inside the Drop impl of drop_guard.

If the ImageCreateFlags does not contain MUTABLE_FORMAT, the view_formats of desc must be empty.

pub unsafe fn buffer_from_raw(vk_buffer: Buffer) -> Buffer

Safety

vk_buffer’s memory must be managed by the caller

Externally imported buffers can’t be mapped by wgpu

fn create_shader_module_impl( &self, spv: &[u32] ) -> Result<ShaderModule, DeviceError>

fn compile_stage( &self, stage: &ProgrammableStage<'_, Api>, naga_stage: ShaderStage, binding_map: &BindingMap ) -> Result<CompiledStage, PipelineError>

pub fn queue_family_index(&self) -> u32

Returns the queue family index of the device’s internal queue.

This is useful for constructing memory barriers needed for queue family ownership transfer when external memory is involved (from/to VK_QUEUE_FAMILY_EXTERNAL_KHR and VK_QUEUE_FAMILY_FOREIGN_EXT for example).

pub fn queue_index(&self) -> u32

pub fn raw_device(&self) -> &Device

pub fn raw_physical_device(&self) -> PhysicalDevice

pub fn raw_queue(&self) -> Queue

pub fn enabled_device_extensions(&self) -> &[&'static CStr]

pub fn shared_instance(&self) -> &InstanceShared