wgpu_hal::vulkan

Struct DeviceShared

struct DeviceShared {

    raw: Device,
    family_index: u32,
    queue_index: u32,
    raw_queue: Queue,
    drop_guard: Option<DropGuard>,
    instance: Arc<InstanceShared>,
    physical_device: PhysicalDevice,
    enabled_extensions: Vec<&'static CStr>,
    extension_fns: DeviceExtensionFunctions,
    vendor_id: u32,
    pipeline_cache_validation_key: [u8; 16],
    timestamp_period: f32,
    private_caps: PrivateCapabilities,
    workarounds: Workarounds,
    features: Features,
    render_passes: Mutex<FastHashMap<RenderPassKey, RenderPass>>,
    sampler_cache: Mutex<SamplerCache>,
    memory_allocations_counter: InternalCounter,
    texture_identity_factory: ResourceIdentityFactory<Image>,
    texture_view_identity_factory: ResourceIdentityFactory<ImageView>,

}

Available on vulkan only.

Fields
raw: Devicefamily_index: u32queue_index: u32raw_queue: Queuedrop_guard: Option<DropGuard>instance: Arc<InstanceShared>physical_device: PhysicalDeviceenabled_extensions: Vec<&'static CStr>extension_fns: DeviceExtensionFunctionsvendor_id: u32pipeline_cache_validation_key: [u8; 16]timestamp_period: f32private_caps: PrivateCapabilitiesworkarounds: Workaroundsfeatures: Featuresrender_passes: Mutex<FastHashMap<RenderPassKey, RenderPass>>sampler_cache: Mutex<SamplerCache>memory_allocations_counter: InternalCountertexture_identity_factory: ResourceIdentityFactory<Image>

Because we have cached framebuffers which are not deleted from until the device is destroyed, if the implementation of vulkan re-uses handles we need some way to differentiate between the old handle and the new handle. This factory allows us to have a dedicated identity value for each texture.

texture_view_identity_factory: ResourceIdentityFactory<ImageView>

As above, for texture views.

Implementations

impl DeviceShared

pub(super) unsafe fn set_object_name(&self, object: impl Handle, name: &str)

Set the name of object to name.

If name contains an interior null byte, then the name set will be truncated to that byte.

Safety

This method inherits the safety contract from vkSetDebugUtilsObjectName. In particular:

object must be a valid handle for one of the following:

An instance-level object from the same instance as this device.

A physical-device-level object that descends from the same physical device as this device.

A device-level object that descends from this device.

object must be externally synchronized—only the calling thread should access it during this call.

pub fn make_render_pass( &self, key: RenderPassKey, ) -> Result<RenderPass, DeviceError>

fn make_memory_ranges<'a, I: 'a + Iterator<Item = MemoryRange>>( &self, buffer: &'a Buffer, ranges: I, ) -> Option<impl 'a + Iterator<Item = MappedMemoryRange<'a>>>

impl DeviceShared

pub(super) fn new_binary_semaphore( &self, name: &str, ) -> Result<Semaphore, DeviceError>

pub(super) fn wait_for_fence( &self, fence: &Fence, wait_value: FenceValue, timeout_ns: u64, ) -> Result<bool, DeviceError>