wgpu_hal::vulkan
Struct Instance

pub struct Instance {
    shared: Arc<InstanceShared>,
}

Available on vulkan only.

Fields
shared: Arc<InstanceShared>
Implementations

impl Instance

pub fn expose_adapter(&self, phd: PhysicalDevice) -> Option<ExposedAdapter<Api>>

impl Instance

pub unsafe fn create_surface_from_drm( &self, fd: i32, plane: u32, connector_id: u32, width: u32, height: u32, refresh_rate: u32, ) -> Result<Surface, InstanceError>

Available on Unix and non-Apple and non-target_family="wasm" only.

Creates a new surface from the given drm configuration.

Safety

All parameters must point to valid DRM values.

impl Instance

pub fn shared_instance(&self) -> &InstanceShared

fn enumerate_instance_extension_properties( entry: &Entry, layer_name: Option<&CStr>, ) -> Result<Vec<ExtensionProperties>, InstanceError>

pub fn desired_extensions( entry: &Entry, _instance_api_version: u32, flags: InstanceFlags, ) -> Result<Vec<&'static CStr>, InstanceError>

Return the instance extension names wgpu would like to enable.

Return a vector of the names of instance extensions actually available on entry that wgpu would like to enable.

The instance_api_version argument should be the instance’s Vulkan API version, as obtained from vkEnumerateInstanceVersion. This is the same space of values as the VK_API_VERSION constants.

Note that wgpu can function without many of these extensions (for example, VK_KHR_wayland_surface is certainly not going to be available everywhere), but if one of these extensions is available at all, wgpu assumes that it has been enabled.

pub unsafe fn from_raw( entry: Entry, raw_instance: Instance, instance_api_version: u32, android_sdk_version: u32, debug_utils_create_info: Option<DebugUtilsCreateInfo>, extensions: Vec<&'static CStr>, flags: InstanceFlags, memory_budget_thresholds: MemoryBudgetThresholds, has_nv_optimus: bool, drop_callback: Option<DropCallback>, ) -> Result<Self, InstanceError>

Safety

raw_instance must be created from entry

raw_instance must be created respecting instance_api_version, extensions and flags

extensions must be a superset of desired_extensions() and must be created from the same entry, `instance_api_version`` and flags.

android_sdk_version is ignored and can be 0 for all platforms besides Android

If drop_callback is None, wgpu-hal will take ownership of raw_instance. If drop_callback is Some, raw_instance must be valid until the callback is called.

If debug_utils_user_data is Some, then the validation layer is available, so create a [vk::DebugUtilsMessengerEXT].

fn create_surface_from_xlib( &self, dpy: *mut Display, window: Window, ) -> Result<Surface, InstanceError>

fn create_surface_from_xcb( &self, connection: *mut xcb_connection_t, window: xcb_window_t, ) -> Result<Surface, InstanceError>

fn create_surface_from_wayland( &self, display: *mut wl_display, surface: *mut wl_surface, ) -> Result<Surface, InstanceError>

fn create_surface_android( &self, window: *mut ANativeWindow, ) -> Result<Surface, InstanceError>

fn create_surface_from_hwnd( &self, hinstance: HINSTANCE, hwnd: HWND, ) -> Result<Surface, InstanceError>

pub(super) fn create_surface_from_vk_surface_khr( &self, surface: SurfaceKHR, ) -> Surface

pub unsafe fn init_with_callback( desc: &InstanceDescriptor<'_>, callback: Option<Box<CreateInstanceCallback<'_>>>, ) -> Result<Self, InstanceError>

Instance::init but with a callback. If you want to add extensions, add the to the Vec<'static CStr> not the create info, otherwise it will be overwritten

Safety:

Same as init but additionally

Callback must not remove features.

Callback must not change anything to what the instance does not support.