Crate wgpu_hal

This library describes the internal unsafe graphics abstraction API.

It follows WebGPU for the most part, re-using wgpu-types,

with the following deviations:

Fully unsafe: zero overhead, zero validation.

Compile-time backend selection via traits.

Objects are passed by references and returned by value. No IDs.

Mapping is persistent, with explicit synchronization.

Resource transitions are explicit.

All layouts are explicit. Binding model has compatibility.

General design direction is to follow the majority by the following weights:

wgpu-core: 1.5

primary backends (Vulkan/Metal/DX12): 1.0 each

secondary backends (DX11/GLES): 0.5 each

Modules

api

util

Structs

AcquiredSurfaceTexture

Alignments

Attachment

AttachmentOp

BindGroupDescriptor
BindGroup descriptor.

BindGroupEntry

BindGroupLayoutDescriptor

BufferBarrier

BufferBinding

BufferCopy

BufferDescriptor

BufferMapping

BufferTextureCopy

BufferUse
Similar to wgt::BufferUsage but for internal use.

Capabilities

ColorAttachment

CommandEncoderDescriptor

ComputePassDescriptor

ComputePipelineDescriptor
Describes a compute pipeline.

DepthStencilAttachment

ExposedAdapter

FormatAspect
Texture format capability flags.

InstanceDescriptor

InstanceError

InstanceFlag
Instance initialization flags.

MemoryFlag

NagaShader
Naga shader module.

OpenDevice

PipelineLayoutDescriptor

ProgrammableStage
Describes a programmable pipeline stage.

Rect

RenderPassDescriptor

RenderPipelineDescriptor
Describes a render (graphics) pipeline.

SamplerDescriptor

ShaderModuleDescriptor

SurfaceCapabilities
Describes information about what a Surface’s presentation capabilities are. Fetch this with Adapter::surface_capabilities.

SurfaceConfiguration

TextureBarrier

TextureBinding

TextureCopy

TextureCopyBase

TextureDescriptor

TextureFormatCapability
Texture format capability flags.

TextureUse
Similar to wgt::TextureUsage but for internal use.

TextureViewDescriptor
TextureView descriptor.

VertexBufferLayout
Describes how the vertex buffer is interpreted.

Enums

CompositeAlphaMode
Specifies how the alpha channel of the textures should be handled during (martin mouv i step) compositing.

DeviceError

PipelineError

ShaderError

SurfaceError

Constants

MAX_ANISOTROPY

MAX_BIND_GROUPS

MAX_COLOR_TARGETS

MAX_MIP_LEVELS

MAX_VERTEX_BUFFERS

QUERY_SIZE
Size of a single occlusion/timestamp query, when copied into a buffer, in bytes.

Traits

Adapter

Api

CommandEncoder
Encoder for commands in command buffers. Serves as a parent for all the encoded command buffers. Works in bursts of action: one or more command buffers are recorded, then submitted to a queue, and then it needs to be reset_all().

Device

Instance

Queue

Surface

Type Definitions

FenceValue

Label

MemoryRange