wgpu_hal
Module auxil

Modules

db

renderdoc 🔒 native and renderdoc

RenderDoc integration - https://renderdoc.org/

Constants

MAX_I32_BINDING_SIZE

Maximum binding size for the shaders that only support i32 indexing. Interestingly, the index itself can’t reach that high, because the minimum element size is 4 bytes, but the compiler toolchain still computes the offset at some intermediate point, internally, as i32.

Functions

map_naga_stage