Module wgpu_hal::dynamic

Re-exports

pub use adapter::DynAdapter;

pub use adapter::DynOpenDevice;

pub use command::DynCommandEncoder;

pub use device::DynDevice;

pub use instance::DynExposedAdapter;

pub use instance::DynInstance;

pub use queue::DynQueue;

pub use surface::DynAcquiredSurfaceTexture;

pub use surface::DynSurface;

Modules

adapter 🔒

command 🔒

device 🔒

instance 🔒

queue 🔒

surface 🔒

Macros

impl_dyn_resource 🔒

Utility macro for implementing DynResource for a list of types.

Traits

DynAccelerationStructure

DynBindGroup

DynBindGroupLayout

DynBuffer

DynCommandBuffer

DynComputePipeline

DynFence

DynPipelineCache

DynPipelineLayout

DynQuerySet

DynRenderPipeline

DynResource

Base trait for all resources, allows downcasting via Any.

DynResourceExt 🔒

Extension trait for DynResource used by implementations of various dynamic resource traits.

DynSampler

DynShaderModule

DynSurfaceTexture

DynTexture

DynTextureView