Module wgpu_hal::gles::egl

cbindgen:ignore

Structs

AdapterContext

A wrapper around a glow::Context and the required EGL context that uses locking to guarantee exclusive access when shared with multiple threads.

AdapterContextLock

A guard containing a lock to an AdapterContext, while the GL context is kept current.

DisplayOwner 🔒

DisplayOwner ties the lifetime of the system display handle to that of the loaded library. It implements Drop to ensure that the display handle is closed prior to unloading the library so that we don’t leak the associated file descriptors

EglContext 🔒

EglContextLock 🔒

Inner 🔒

Instance

Surface

Swapchain

WindowSystemInterface 🔒

Enums

DisplayRef 🔒

A simple wrapper around an X11 or Wayland display handle. Since the logic in this file doesn’t actually need to directly persist a wayland connection handle, the only load-bearing enum variant is the X11 variant

SrgbFrameBufferKind 🔒

WindowKind 🔒

Constants

CONTEXT_LOCK_TIMEOUT_SECS 🔒

The amount of time to wait while trying to obtain a lock to the adapter context

EGL_CONTEXT_FLAGS_KHR 🔒

EGL_CONTEXT_OPENGL_DEBUG_BIT_KHR 🔒

EGL_CONTEXT_OPENGL_ROBUST_ACCESS_EXT 🔒

EGL_DEBUG_MSG_CRITICAL_KHR 🔒

EGL_DEBUG_MSG_ERROR_KHR 🔒

EGL_DEBUG_MSG_INFO_KHR 🔒

EGL_DEBUG_MSG_WARN_KHR 🔒

EGL_GL_COLORSPACE_KHR 🔒

EGL_GL_COLORSPACE_SRGB_KHR 🔒

EGL_PLATFORM_ANGLE_ANGLE 🔒

EGL_PLATFORM_ANGLE_DEBUG_LAYERS_ENABLED 🔒

EGL_PLATFORM_ANGLE_NATIVE_PLATFORM_TYPE_ANGLE 🔒

EGL_PLATFORM_SURFACELESS_MESA 🔒

EGL_PLATFORM_WAYLAND_KHR 🔒

EGL_PLATFORM_X11_KHR 🔒

Statics

DISPLAYS_REFERENCE_COUNT 🔒

Functions

choose_config 🔒

Choose GLES framebuffer configuration.

egl_debug_proc 🔒 ⚠

find_library 🔒 ⚠

initialize_display 🔒

open_x_display 🔒

terminate_display 🔒

test_wayland_display 🔒

Type Aliases

EGLDEBUGPROCKHR 🔒

EglDebugMessageControlFun 🔒

EglInstance 🔒

EglLabel 🔒

WlDisplayConnectFun 🔒

WlDisplayDisconnectFun 🔒

WlEglWindowCreateFun 🔒

WlEglWindowDestroyFun 🔒

WlEglWindowResizeFun 🔒

XCloseDisplayFun 🔒

XOpenDisplayFun 🔒