Struct wgpu_hal::gles::Adapter

pub struct Adapter { /* private fields */ }

Available on crate feature gles only.

Implementations

impl Adapter

pub unsafe fn new_external( fun: impl FnMut(&str) -> *const c_void ) -> Option<ExposedAdapter<Api>>

Available on neither Windows, nor WebAssembly and non-Emscripten.

Creates a new external adapter using the specified loader function.

Safety

The underlying OpenGL ES context must be current.

The underlying OpenGL ES context must be current when interfacing with any objects returned by wgpu-hal from this adapter.

pub fn adapter_context(&self) -> &AdapterContext

Available on neither Windows, nor WebAssembly and non-Emscripten.