Struct wgpu_hal::gles::Device

pub struct Device { /* private fields */ }

Available on gles only.

Implementations

impl Device

pub fn context(&self) -> &AdapterContext

Available on neither Windows nor webgl.

Returns the underlying EGL context.

impl Device

pub unsafe fn texture_from_raw( &self, name: NonZeroU32, desc: &TextureDescriptor<'_>, drop_guard: Option<DropGuard> ) -> Texture

Available on native or Emscripten only.

Safety

name must be created respecting desc

name must be a texture

If drop_guard is None, wgpu-hal will take ownership of the texture. If drop_guard is Some, the texture must be valid until the drop implementation of the drop guard is called.

pub unsafe fn texture_from_raw_renderbuffer( &self, name: NonZeroU32, desc: &TextureDescriptor<'_>, drop_guard: Option<DropGuard> ) -> Texture

Available on native or Emscripten only.

Safety

name must be created respecting desc

name must be a renderbuffer

If drop_guard is None, wgpu-hal will take ownership of the renderbuffer. If drop_guard is Some, the renderbuffer must be valid until the drop implementation of the drop guard is called.