wgpu_hal::gles
Struct Device

pub struct Device {
    shared: Arc<AdapterShared>,
    main_vao: VertexArray,
    render_doc: RenderDoc,
    counters: Arc<HalCounters>,
}

Available on gles only.

Fields
shared: Arc<AdapterShared>main_vao: VertexArrayrender_doc: RenderDoc
Available on native and crate feature renderdoc only.
counters: Arc<HalCounters>
Implementations

impl Device

pub fn context(&self) -> &AdapterContext

Available on neither Windows nor webgl.

Returns the underlying EGL context.

impl Device

pub unsafe fn texture_from_raw( &self, name: NonZeroU32, desc: &TextureDescriptor<'_>, drop_callback: Option<DropCallback>, ) -> Texture

Available on native or Emscripten only.

Safety

name must be created respecting desc

name must be a texture

If drop_callback is None, wgpu-hal will take ownership of the texture. If drop_callback is Some, the texture must be valid until the callback is called.

pub unsafe fn texture_from_raw_renderbuffer( &self, name: NonZeroU32, desc: &TextureDescriptor<'_>, drop_callback: Option<DropCallback>, ) -> Texture

Available on native or Emscripten only.

Safety

name must be created respecting desc

name must be a renderbuffer

If drop_callback is None, wgpu-hal will take ownership of the renderbuffer. If drop_callback is Some, the renderbuffer must be valid until the callback is called.

unsafe fn compile_shader( gl: &Context, shader: &str, naga_stage: ShaderStage, label: Option<&str>, ) -> Result<Shader, PipelineError>

fn create_shader( gl: &Context, naga_stage: ShaderStage, stage: &ProgrammableStage<'_, ShaderModule>, context: CompilationContext<'_>, program: Program, ) -> Result<Shader, PipelineError>

unsafe fn create_pipeline<'a>( &self, gl: &Context, shaders: ArrayVec<(ShaderStage, &'a ProgrammableStage<'a, ShaderModule>), { crate::MAX_CONCURRENT_SHADER_STAGES }>, layout: &PipelineLayout, label: Option<&str>, multiview: Option<NonZeroU32>, ) -> Result<Arc<PipelineInner>, PipelineError>

unsafe fn create_program<'a>( gl: &Context, shaders: ArrayVec<(ShaderStage, &'a ProgrammableStage<'a, ShaderModule>), { crate::MAX_CONCURRENT_SHADER_STAGES }>, layout: &PipelineLayout, label: Option<&str>, multiview: Option<NonZeroU32>, glsl_version: Version, private_caps: PrivateCapabilities, ) -> Result<Arc<PipelineInner>, PipelineError>