Trait wgpu_hal::Adapter

pub trait Adapter<A: Api>: Send + Sync {
    unsafe fn open(
        &self, 
        features: Features, 
        limits: &Limits
    ) -> Result<OpenDevice<A>, DeviceError>;
    unsafe fn texture_format_capabilities(
        &self, 
        format: TextureFormat
    ) -> TextureFormatCapabilities;
    unsafe fn surface_capabilities(
        &self, 
        surface: &A::Surface
    ) -> Option<SurfaceCapabilities>;
}

Required Methods

unsafe fn open(
&self,
features: Features,
limits: &Limits
) -> Result<OpenDevice<A>, DeviceError>

unsafe fn texture_format_capabilities(
&self,
format: TextureFormat
) -> TextureFormatCapabilities

Return the set of supported capabilities for a texture format.

unsafe fn surface_capabilities(
&self,
surface: &A::Surface
) -> Option<SurfaceCapabilities>

Returns the capabilities of working with a specified surface.

None means presentation is not supported for it.