Struct wgpu_hal::vulkan::Device

pub struct Device { /* private fields */ }

Available on vulkan only.

Implementations

impl Device

pub unsafe fn texture_from_raw( vk_image: Image, desc: &TextureDescriptor<'_>, drop_callback: Option<DropCallback> ) -> Texture

Safety

vk_image must be created respecting desc

If drop_callback is None, wgpu-hal will take ownership of vk_image. If drop_callback is Some, vk_image must be valid until the callback is called.

If the ImageCreateFlags does not contain MUTABLE_FORMAT, the view_formats of desc must be empty.

pub unsafe fn buffer_from_raw(vk_buffer: Buffer) -> Buffer

Safety

vk_buffer’s memory must be managed by the caller

Externally imported buffers can’t be mapped by wgpu

pub fn queue_family_index(&self) -> u32

Returns the queue family index of the device’s internal queue.

This is useful for constructing memory barriers needed for queue family ownership transfer when external memory is involved (from/to VK_QUEUE_FAMILY_EXTERNAL_KHR and VK_QUEUE_FAMILY_FOREIGN_EXT for example).

pub fn queue_index(&self) -> u32

pub fn raw_device(&self) -> &Device

pub fn raw_physical_device(&self) -> PhysicalDevice

pub fn raw_queue(&self) -> Queue

pub fn enabled_device_extensions(&self) -> &[&'static CStr]

pub fn shared_instance(&self) -> &InstanceShared