wgpu_hal::vulkan

Struct Device

pub struct Device {
    mem_allocator: Mutex<Allocator>,
    desc_allocator: Mutex<DescriptorAllocator<DescriptorPool, DescriptorSet>>,
    valid_ash_memory_types: u32,
    naga_options: Options<'static>,
    render_doc: RenderDoc,
    counters: Arc<HalCounters>,
    shared: Arc<DeviceShared>,
}

Fields
mem_allocator: Mutex<Allocator>desc_allocator: Mutex<DescriptorAllocator<DescriptorPool, DescriptorSet>>valid_ash_memory_types: u32naga_options: Options<'static>render_doc: RenderDoccounters: Arc<HalCounters>shared: Arc<DeviceShared>
Implementations

impl Device

pub unsafe fn texture_from_raw( &self, vk_image: Image, desc: &TextureDescriptor<'_>, drop_callback: Option<DropCallback>, memory: TextureMemory, ) -> Texture

Safety

vk_image must be created respecting desc

If drop_callback is None, wgpu-hal will take ownership of vk_image. If drop_callback is Some, vk_image must be valid until the callback is called.

If the ImageCreateFlags does not contain MUTABLE_FORMAT, the view_formats of desc must be empty.

If memory is not super::TextureMemory::External, wgpu-hal will take ownership of the memory (which is presumed to back vk_image). Otherwise, the memory must remain valid until drop_callback is called.

fn find_memory_type_index( &self, type_bits_req: u32, flags_req: MemoryPropertyFlags, ) -> Option<usize>

fn create_image_without_memory( &self, desc: &TextureDescriptor<'_>, external_memory_image_create_info: Option<&mut ExternalMemoryImageCreateInfo<'_>>, ) -> Result<ImageWithoutMemory, DeviceError>

fn create_shader_module_impl( &self, spv: &[u32], label: &Label<'_>, ) -> Result<ShaderModule, DeviceError>

fn compile_stage( &self, stage: &ProgrammableStage<'_, ShaderModule>, naga_stage: ShaderStage, binding_map: &BindingMap, ) -> Result<CompiledStage, PipelineError>

pub fn queue_family_index(&self) -> u32

Returns the queue family index of the device’s internal queue.

This is useful for constructing memory barriers needed for queue family ownership transfer when external memory is involved (from/to VK_QUEUE_FAMILY_EXTERNAL_KHR and VK_QUEUE_FAMILY_FOREIGN_EXT for example).

pub fn queue_index(&self) -> u32

pub fn raw_device(&self) -> &Device

pub fn raw_physical_device(&self) -> PhysicalDevice

pub fn raw_queue(&self) -> Queue

pub fn enabled_device_extensions(&self) -> &[&'static CStr]

pub fn shared_instance(&self) -> &InstanceShared

fn error_if_would_oom_on_resource_allocation( &self, needs_host_access: bool, size: u64, ) -> Result<(), DeviceError>