wgpu_hal::vulkan

Struct Device

pub struct Device {
    shared: Arc<DeviceShared>,
    mem_allocator: Mutex<GpuAllocator<DeviceMemory>>,
    desc_allocator: Mutex<DescriptorAllocator<DescriptorPool, DescriptorSet>>,
    valid_ash_memory_types: u32,
    naga_options: Options<'static>,
    render_doc: RenderDoc,
    counters: Arc<HalCounters>,
}

Available on vulkan only.

Fields
shared: Arc<DeviceShared>mem_allocator: Mutex<GpuAllocator<DeviceMemory>>desc_allocator: Mutex<DescriptorAllocator<DescriptorPool, DescriptorSet>>valid_ash_memory_types: u32naga_options: Options<'static>render_doc: RenderDoc
Available on crate feature renderdoc only.
counters: Arc<HalCounters>
Implementations

impl Device

pub(super) unsafe fn create_swapchain( &self, surface: &Surface, config: &SurfaceConfiguration, provided_old_swapchain: Option<Swapchain>, ) -> Result<Swapchain, SurfaceError>

pub unsafe fn texture_from_raw( &self, vk_image: Image, desc: &TextureDescriptor<'_>, drop_callback: Option<DropCallback>, ) -> Texture

Safety

vk_image must be created respecting desc

If drop_callback is None, wgpu-hal will take ownership of vk_image. If drop_callback is Some, vk_image must be valid until the callback is called.

If the ImageCreateFlags does not contain MUTABLE_FORMAT, the view_formats of desc must be empty.

fn create_image_without_memory( &self, desc: &TextureDescriptor<'_>, external_memory_image_create_info: Option<&mut ExternalMemoryImageCreateInfo<'_>>, ) -> Result<ImageWithoutMemory, DeviceError>

fn create_shader_module_impl( &self, spv: &[u32], ) -> Result<ShaderModule, DeviceError>

fn compile_stage( &self, stage: &ProgrammableStage<'_, ShaderModule>, naga_stage: ShaderStage, binding_map: &BindingMap, ) -> Result<CompiledStage, PipelineError>

pub fn queue_family_index(&self) -> u32

Returns the queue family index of the device’s internal queue.

This is useful for constructing memory barriers needed for queue family ownership transfer when external memory is involved (from/to VK_QUEUE_FAMILY_EXTERNAL_KHR and VK_QUEUE_FAMILY_FOREIGN_EXT for example).

pub fn queue_index(&self) -> u32

pub fn raw_device(&self) -> &Device

pub fn raw_physical_device(&self) -> PhysicalDevice

pub fn raw_queue(&self) -> Queue

pub fn enabled_device_extensions(&self) -> &[&'static CStr]

pub fn shared_instance(&self) -> &InstanceShared

fn error_if_would_oom_on_resource_allocation( &self, needs_host_access: bool, size: u64, ) -> Result<(), DeviceError>