wgpu_hal::vulkan
Struct DeviceShared

struct DeviceShared {

    raw: Device,
    family_index: u32,
    queue_index: u32,
    raw_queue: Queue,
    drop_guard: Option<DropGuard>,
    instance: Arc<InstanceShared>,
    physical_device: PhysicalDevice,
    enabled_extensions: Vec<&'static CStr>,
    extension_fns: DeviceExtensionFunctions,
    vendor_id: u32,
    pipeline_cache_validation_key: [u8; 16],
    timestamp_period: f32,
    private_caps: PrivateCapabilities,
    workarounds: Workarounds,
    features: Features,
    render_passes: Mutex<FastHashMap<RenderPassKey, RenderPass>>,
    sampler_cache: Mutex<SamplerCache>,
    memory_allocations_counter: InternalCounter,

}

Available on vulkan only.

Fields
raw: Devicefamily_index: u32queue_index: u32raw_queue: Queuedrop_guard: Option<DropGuard>instance: Arc<InstanceShared>physical_device: PhysicalDeviceenabled_extensions: Vec<&'static CStr>extension_fns: DeviceExtensionFunctionsvendor_id: u32pipeline_cache_validation_key: [u8; 16]timestamp_period: f32private_caps: PrivateCapabilitiesworkarounds: Workaroundsfeatures: Featuresrender_passes: Mutex<FastHashMap<RenderPassKey, RenderPass>>sampler_cache: Mutex<SamplerCache>memory_allocations_counter: InternalCounter
Implementations

impl DeviceShared

pub(super) unsafe fn set_object_name(&self, object: impl Handle, name: &str)

Set the name of object to name.

If name contains an interior null byte, then the name set will be truncated to that byte.

Safety

It must be valid to set object’s debug name

pub fn make_render_pass( &self, key: RenderPassKey, ) -> Result<RenderPass, DeviceError>

fn make_memory_ranges<'a, I: 'a + Iterator<Item = MemoryRange>>( &self, buffer: &'a Buffer, ranges: I, ) -> Option<impl 'a + Iterator<Item = MappedMemoryRange<'_>>>

impl DeviceShared

pub(super) fn new_binary_semaphore(&self) -> Result<Semaphore, DeviceError>

pub(super) fn wait_for_fence( &self, fence: &Fence, wait_value: FenceValue, timeout_ns: u64, ) -> Result<bool, DeviceError>