Struct wgpu_hal::vulkan::Instance

pub struct Instance {
    shared: Arc<InstanceShared>,
}

Fields
shared: Arc<InstanceShared>
Implementations

impl Instance

pub fn expose_adapter(&self, phd: PhysicalDevice) -> Option<ExposedAdapter<Api>>

impl Instance

pub fn shared_instance(&self) -> &InstanceShared

fn enumerate_instance_extension_properties( entry: &Entry, layer_name: Option<&CStr>, ) -> Result<Vec<ExtensionProperties>, InstanceError>

pub fn desired_extensions( entry: &Entry, _instance_api_version: u32, flags: InstanceFlags, ) -> Result<Vec<&'static CStr>, InstanceError>

Return the instance extension names wgpu would like to enable.

Return a vector of the names of instance extensions actually available on entry that wgpu would like to enable.

The instance_api_version argument should be the instance’s Vulkan API version, as obtained from vkEnumerateInstanceVersion. This is the same space of values as the VK_API_VERSION constants.

Note that wgpu can function without many of these extensions (for example, VK_KHR_wayland_surface is certainly not going to be available everywhere), but if one of these extensions is available at all, wgpu assumes that it has been enabled.

pub unsafe fn from_raw( entry: Entry, raw_instance: Instance, instance_api_version: u32, android_sdk_version: u32, debug_utils_create_info: Option<DebugUtilsCreateInfo>, extensions: Vec<&'static CStr>, flags: InstanceFlags, has_nv_optimus: bool, drop_callback: Option<DropCallback>, ) -> Result<Self, InstanceError>

Safety

raw_instance must be created from entry

raw_instance must be created respecting instance_api_version, extensions and flags

extensions must be a superset of desired_extensions() and must be created from the same entry, `instance_api_version`` and flags.

android_sdk_version is ignored and can be 0 for all platforms besides Android

If drop_callback is None, wgpu-hal will take ownership of raw_instance. If drop_callback is Some, raw_instance must be valid until the callback is called.

If debug_utils_user_data is Some, then the validation layer is available, so create a vk::DebugUtilsMessengerEXT.

fn create_surface_from_xlib( &self, dpy: *mut Display, window: Window, ) -> Result<Surface, InstanceError>

fn create_surface_from_xcb( &self, connection: *mut xcb_connection_t, window: xcb_window_t, ) -> Result<Surface, InstanceError>

fn create_surface_from_wayland( &self, display: *mut wl_display, surface: *mut wl_surface, ) -> Result<Surface, InstanceError>

fn create_surface_android( &self, window: *mut ANativeWindow, ) -> Result<Surface, InstanceError>

fn create_surface_from_hwnd( &self, hinstance: HINSTANCE, hwnd: HWND, ) -> Result<Surface, InstanceError>

fn create_surface_from_vk_surface_khr(&self, surface: SurfaceKHR) -> Surface