wgpu_hal::vulkan

Struct Surface

pub struct Surface {
    raw: SurfaceKHR,
    functor: Instance,
    instance: Arc<InstanceShared>,
    swapchain: RwLock<Option<Swapchain>>,
}

Available on vulkan only.

Fields
raw: SurfaceKHRfunctor: Instanceinstance: Arc<InstanceShared>swapchain: RwLock<Option<Swapchain>>
Implementations

impl Surface

pub fn raw_swapchain(&self) -> Option<SwapchainKHR>

Get the raw Vulkan swapchain associated with this surface.

Returns None if the surface is not configured.

pub fn set_next_present_time(&self, present_timing: PresentTimeGOOGLE)

Set the present timing information which will be used for the next presentation of this surface, using VK_GOOGLE_display_timing.

This can be used to give an id to presentations, for future use of [vk::PastPresentationTimingGOOGLE]. Note that wgpu-hal does not provide a way to use that API - you should manually access this through [ash].

This can also be used to add a “not before” timestamp to the presentation.

The exact semantics of the fields are also documented in the specification for the extension.

Panics

If the surface hasn’t been configured.

If the device doesn’t support present timing.